refactor + demo
