+++ nama/__init__.py
@@ -1,2 +1,4 @@
 from .matcher import Matcher
 from .utils import *
+from .models.similarity_model import SimilarityModel
+from .scoring import score_predicted
